refactor(runner): code governance — DIP fix, file splits, dead code cleanup, test coverage
- Fix dead code: simplify redundant if/else in runner_mcp.go
- Add //go:build integration tag to 6 integration test files
- Extract MessageHandlerContext interface for DIP compliance
- Split 13 large files (>300 lines) into focused modules
- Add client package tests: coverage 10.7% → 50.1%
