+++ runner/internal/autopilot/autopilot_controller.go
@@ -213,170 +213,5 @@ func (ac *AutopilotController) GetStatus() Status {
 	}
 }
 
-// OnPodWaiting is called when the Pod transitions to waiting state.
-// This is the main event-driven entry point triggered by StateDetectorCoordinator.
-// Includes deduplication to prevent rapid re-triggering.
-func (ac *AutopilotController) OnPodWaiting() {
-	ac.log.Debug("OnPodWaiting triggered", "autopilot_key", ac.key)
-
-	// Check trigger deduplication
-	if !ac.iterCtrl.CheckTriggerDedup() {
-		ac.log.Debug("Skipping iteration - deduplication", "autopilot_key", ac.key)
-		return
-	}
-
-	// Check if user has taken over
-	if ac.userHandler.IsUserTakeover() {
-		ac.log.Debug("Skipping iteration - user takeover", "autopilot_key", ac.key)
-		return
-	}
-
-	// Check if phase allows iteration
-	if !ac.phaseMgr.CanProcessIteration() {
-		ac.log.Debug("Skipping iteration - phase not ready", "autopilot_key", ac.key, "phase", ac.phaseMgr.GetPhase())
-		return
-	}
-
-	// Check max iterations (the only hard protection)
-	if ac.iterCtrl.HasReachedMaxIterations() {
-		ac.phaseMgr.SetPhase(PhaseMaxIterations)
-		ac.log.Info("Max iterations reached", "autopilot_key", ac.key)
-		if ac.reporter != nil {
-			ac.reporter.ReportAutopilotTerminated(&runnerv1.AutopilotTerminatedEvent{
-				AutopilotKey: ac.key,
-				Reason:       "max_iterations",
-			})
-		}
-		return
-	}
-
-	// Increment iteration
-	iteration, ok := ac.iterCtrl.IncrementIteration()
-	if !ok {
-		// Max iterations reached during increment
-		ac.phaseMgr.SetPhase(PhaseMaxIterations)
-		ac.log.Info("Max iterations reached", "autopilot_key", ac.key)
-		if ac.reporter != nil {
-			ac.reporter.ReportAutopilotTerminated(&runnerv1.AutopilotTerminatedEvent{
-				AutopilotKey: ac.key,
-				Reason:       "max_iterations",
-			})
-		}
-		return
-	}
-
-	// Report iteration started
-	ac.iterCtrl.ReportIterationEvent(iteration, "started", "", nil)
-
-	// Run single decision in a goroutine.
-	// Acquire wgMu to ensure atomicity of the stopped check + wg.Add(1).
-	// Without this lock, Stop() could set stopped=true and call wg.Wait()
-	// between our ctx check and wg.Add(1), causing a panic.
-	ac.wgMu.Lock()
-	if ac.stopped {
-		ac.wgMu.Unlock()
-		ac.log.Debug("Controller stopped before starting decision goroutine", "autopilot_key", ac.key)
-		return
-	}
-	ac.wg.Add(1)
-	ac.wgMu.Unlock()
-
-	go func() {
-		defer ac.wg.Done()
-		ac.runSingleDecision(iteration)
-	}()
-}
-
-// sendInitialPrompt starts the first iteration when Pod is waiting.
-// This launches the control process which will use MCP tools to interact with Pod.
-func (ac *AutopilotController) sendInitialPrompt() {
-	ac.log.Info("Starting initial iteration", "autopilot_key", ac.key)
-
-	// Update trigger time to prevent OnPodWaiting from double-triggering
-	ac.iterCtrl.UpdateTriggerTime()
-
-	// Set initial iteration
-	iteration := ac.iterCtrl.SetInitialIteration()
-
-	// Report iteration started
-	ac.iterCtrl.ReportIterationEvent(iteration, "started", "", nil)
-
-	// Run the control process.
-	// Acquire wgMu to ensure atomicity of the stopped check + wg.Add(1).
-	ac.wgMu.Lock()
-	if ac.stopped {
-		ac.wgMu.Unlock()
-		ac.log.Debug("Controller stopped before starting initial decision goroutine", "autopilot_key", ac.key)
-		return
-	}
-	ac.wg.Add(1)
-	ac.wgMu.Unlock()
-
-	go func() {
-		defer ac.wg.Done()
-		ac.runSingleDecision(iteration)
-	}()
-}
-
-// =============================================================================
-// Test Helper Methods (unexported)
-// These methods provide access to internal components for testing purposes.
-// They delegate to the appropriate component and are not part of the public API.
-// =============================================================================
-
-func (ac *AutopilotController) extractSessionID(output string) {
-	if sessionID := ExtractSessionID(output); sessionID != "" {
-		ac.controlRunner.SetSessionID(sessionID)
-	}
-}
-
-func (ac *AutopilotController) getSessionID() string {
-	return ac.controlRunner.GetSessionID()
-}
-
-func (ac *AutopilotController) buildInitialPrompt() string {
-	return ac.promptBuilder.BuildInitialPrompt()
-}
-
-func (ac *AutopilotController) buildResumePrompt(iteration int) string {
-	return ac.promptBuilder.BuildResumePrompt(iteration)
-}
-
-func (ac *AutopilotController) parseDecision(output string) *ControlDecision {
-	return NewDecisionParser().ParseDecision(output)
-}
-
-func (ac *AutopilotController) setPhaseForTest(phase Phase) {
-	ac.phaseMgr.SetPhaseWithoutReport(phase)
-}
-
-func (ac *AutopilotController) setIterationForTest(iteration int) {
-	ac.iterCtrl.mu.Lock()
-	ac.iterCtrl.currentIter = iteration
-	ac.iterCtrl.mu.Unlock()
-}
-
-// GetProgressSummary returns a human-readable summary of task progress.
-func (ac *AutopilotController) GetProgressSummary() string {
-	if ac.progressTracker == nil {
-		return "No progress tracking available"
-	}
-	return ac.progressTracker.GenerateSummary()
-}
-
-// IsStuck checks if no progress has been made for the specified duration.
-func (ac *AutopilotController) IsStuck(threshold time.Duration) bool {
-	if ac.progressTracker == nil {
-		return false
-	}
-	return ac.progressTracker.IsStuck(threshold)
-}
-
-// GetChangedFiles returns files changed during the autopilot session.
-func (ac *AutopilotController) GetChangedFiles() []string {
-	if ac.progressTracker == nil {
-		return nil
-	}
-	startedAt := ac.iterCtrl.GetStartedAt()
-	return ac.progressTracker.GetChangedFilesSince(startedAt)
-}
+// Note: OnPodWaiting, sendInitialPrompt are in autopilot_controller_logic.go
+// Note: Test helpers and progress methods are in autopilot_controller_helpers.go
