+++ runner/internal/client/grpc_connection.go
@@ -3,23 +3,16 @@ package client
 
 import (
 	"fmt"
-	"net"
-	"net/url"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
 	runnerv1 "github.com/anthropics/agentsmesh/proto/gen/go/runner/v1"
 	"github.com/anthropics/agentsmesh/runner/internal/logger"
-	"github.com/anthropics/agentsmesh/runner/internal/safego"
 	"golang.org/x/time/rate"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/credentials/tls/certprovider"
-	"google.golang.org/grpc/keepalive"
-	"google.golang.org/grpc/status"
 )
 
 // GRPCProtocolVersion is the current gRPC protocol version.
@@ -184,237 +177,7 @@ func (c *GRPCConnection) GetOrgSlug() string {
 	return c.orgSlug
 }
 
-// IsInitialized returns whether the connection has completed initialization.
-func (c *GRPCConnection) IsInitialized() bool {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	return c.initialized
-}
-
-// GetAvailableAgents returns the list of available agents.
-func (c *GRPCConnection) GetAvailableAgents() []string {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	return c.availableAgents
-}
-
-// Connect establishes a gRPC connection with mTLS using advancedtls for certificate hot-reloading.
-func (c *GRPCConnection) Connect() error {
-	// Close existing connection and certificate providers (important for reconnection)
-	// This prevents resource leaks, TLS session conflicts, and goroutine leaks from pemfile providers
-	c.mu.Lock()
-	if c.conn != nil {
-		logger.GRPC().Debug("Closing existing gRPC connection before reconnect")
-		c.conn.Close()
-		c.conn = nil
-	}
-	// Close old certificate providers to prevent goroutine leaks
-	// Each pemfile.Provider runs background goroutines for file watching
-	if c.identityProvider != nil {
-		c.identityProvider.Close()
-		c.identityProvider = nil
-	}
-	if c.rootProvider != nil {
-		c.rootProvider.Close()
-		c.rootProvider = nil
-	}
-	c.mu.Unlock()
-
-	// Parse endpoint to extract host:port (remove scheme like grpcs://)
-	dialTarget, err := parseGRPCEndpoint(c.endpoint)
-	if err != nil {
-		return fmt.Errorf("failed to parse gRPC endpoint: %w", err)
-	}
-
-	// Create advancedtls credentials with file-based certificate reloading
-	creds, err := c.createAdvancedTLSCredentials()
-	if err != nil {
-		return fmt.Errorf("failed to create TLS credentials: %w", err)
-	}
-
-	// Fix advancedtls SNI bug: advancedtls.ClientHandshake sets ServerName to the
-	// full authority string (e.g. "agentsmesh.ai:9443" including port), unlike
-	// standard gRPC credentials which strips the port via net.SplitHostPort.
-	// An SNI with port violates RFC 6066 and is rejected by some network middleboxes
-	// (corporate firewalls, DPI). Set the correct hostname-only ServerName here so
-	// advancedtls won't overwrite it with the authority.
-	host, _, err := net.SplitHostPort(dialTarget)
-	if err != nil {
-		host = dialTarget // No port in dialTarget, use as-is
-	}
-	if err := creds.OverrideServerName(host); err != nil { //nolint:staticcheck // advancedtls has no other public API to set ServerName; the unexported serverNameOverride field is inaccessible
-		logger.GRPC().Warn("Failed to override TLS server name", "error", err)
-	}
-
-	// gRPC dial options
-	dialOpts := []grpc.DialOption{
-		grpc.WithTransportCredentials(creds),
-		grpc.WithKeepaliveParams(keepalive.ClientParameters{
-			Time:                30 * time.Second,
-			Timeout:             10 * time.Second,
-			PermitWithoutStream: true,
-		}),
-	}
-
-	// Connect to server
-	conn, err := grpc.Dial(dialTarget, dialOpts...)
-	if err != nil {
-		return fmt.Errorf("failed to dial gRPC server: %w", err)
-	}
-
-	// Create gRPC service client
-	client := runnerv1.NewRunnerServiceClient(conn)
-
-	c.mu.Lock()
-	c.conn = conn
-	c.client = client
-	c.creds = creds
-	c.initialized = false
-	c.mu.Unlock()
-
-	logger.GRPC().Info("Connected to server with advancedtls", "endpoint", c.endpoint, "org", c.orgSlug)
-	return nil
-}
-
-// Start starts the connection management loop.
-func (c *GRPCConnection) Start() {
-	logger.GRPC().Info("gRPC connection manager starting", "endpoint", c.endpoint)
-	c.loopWg.Add(1)
-	safego.Go("grpc-connection-loop", func() {
-		defer c.loopWg.Done()
-		c.connectionLoop()
-	})
-}
-
-// Stop stops the connection and releases resources.
-func (c *GRPCConnection) Stop() {
-	c.stopOnce.Do(func() {
-		logger.GRPC().Info("gRPC connection stopping")
-		close(c.stopCh)
-
-		// Wait for connectionLoop to exit before cleaning up resources
-		c.loopWg.Wait()
-
-		// Wait for in-flight async handlers (handleCreatePod, etc.)
-		c.handlerWg.Wait()
-
-		c.mu.Lock()
-		if c.conn != nil {
-			c.conn.Close()
-			c.conn = nil
-		}
-		// Close certificate providers to prevent goroutine leaks
-		if c.identityProvider != nil {
-			c.identityProvider.Close()
-			c.identityProvider = nil
-		}
-		if c.rootProvider != nil {
-			c.rootProvider.Close()
-			c.rootProvider = nil
-		}
-		c.mu.Unlock()
-		logger.GRPC().Info("gRPC connection stopped")
-	})
-}
-
+// Note: Connect, Start, Stop are in grpc_connection_connect.go
 // Note: connectionLoop and runConnection are in grpc_connection_loop.go
-
-// ==================== gRPC Error Handling ====================
-
-// setFatalError records a fatal error that should stop reconnection attempts.
-func (c *GRPCConnection) setFatalError(err error) {
-	c.fatalErrMu.Lock()
-	c.fatalErr = err
-	c.fatalErrMu.Unlock()
-}
-
-// getFatalError returns the fatal error if one has been recorded.
-func (c *GRPCConnection) getFatalError() error {
-	c.fatalErrMu.Lock()
-	defer c.fatalErrMu.Unlock()
-	return c.fatalErr
-}
-
-// isFatalStreamError checks if a gRPC stream error is fatal (should not retry).
-// Returns true and a user-friendly message if the error is fatal.
-func isFatalStreamError(err error) (bool, string) {
-	st, ok := status.FromError(err)
-	if !ok {
-		return false, ""
-	}
-
-	switch st.Code() {
-	case codes.Unauthenticated:
-		msg := st.Message()
-		if strings.Contains(msg, "runner not found") {
-			return true, "This runner has been deleted from the server. Please re-register with: agentsmesh-runner register --server <SERVER> --token <TOKEN> --force"
-		}
-		return true, "Authentication failed: " + msg + ". Please re-register with: agentsmesh-runner register --server <SERVER> --token <TOKEN> --force"
-
-	case codes.PermissionDenied:
-		msg := st.Message()
-		if strings.Contains(msg, "runner is disabled") {
-			return true, "This runner has been disabled by an administrator. Please contact your organization admin to re-enable it."
-		}
-		return true, "Permission denied: " + msg
-
-	default:
-		return false, ""
-	}
-}
-
-// LastActivityTime returns the most recent send or recv timestamp.
-// Used by the Watchdog health checker to detect stuck connections.
-func (c *GRPCConnection) LastActivityTime() time.Time {
-	sendNs := c.lastSendTime.Load()
-	recvNs := c.lastRecvTime.Load()
-	latest := max(sendNs, recvNs)
-	if latest == 0 {
-		return time.Time{}
-	}
-	return time.Unix(0, latest)
-}
-
-// Ensure GRPCConnection implements Connection interface.
-var _ Connection = (*GRPCConnection)(nil)
-
-// ==================== Helper Functions ====================
-
-// parseGRPCEndpoint parses a gRPC endpoint URL and returns the host:port for dialing.
-// Supports formats:
-//   - grpcs://host:port -> host:port (TLS)
-//   - grpc://host:port  -> host:port (plain)
-//   - host:port         -> host:port (as-is)
-func parseGRPCEndpoint(endpoint string) (string, error) {
-	log := logger.GRPC()
-
-	// If it doesn't contain a scheme, assume it's already host:port
-	if !strings.Contains(endpoint, "://") {
-		return endpoint, nil
-	}
-
-	// Parse as URL
-	u, err := url.Parse(endpoint)
-	if err != nil {
-		log.Error("Invalid endpoint URL", "endpoint", endpoint, "error", err)
-		return "", err
-	}
-
-	// Validate scheme
-	switch u.Scheme {
-	case "grpc", "grpcs":
-		// Valid gRPC schemes
-	default:
-		log.Error("Unsupported gRPC scheme", "scheme", u.Scheme, "endpoint", endpoint)
-		return "", fmt.Errorf("unsupported scheme %q", u.Scheme)
-	}
-
-	// Return host:port
-	if u.Host == "" {
-		log.Error("Missing host in endpoint URL", "endpoint", endpoint)
-		return "", fmt.Errorf("missing host in endpoint")
-	}
-
-	logger.GRPCTrace().Trace("Parsed gRPC endpoint", "endpoint", endpoint, "dial_target", u.Host)
-	return u.Host, nil
-}
+// Note: State query methods are in grpc_connection_state.go
+// Note: Error handling methods are in grpc_connection_error.go
