+++ runner/internal/mcp/http_tools_format_integration_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package mcp
 
 import (
