+++ runner/internal/mcp/mcp_integration_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package mcp
 
 import (
