+++ runner/internal/runner/message_handler_integration_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package runner
 
 import (
