+++ runner/internal/runner/pod_builder_build.go
@@ -1,0 +1,243 @@
+package runner
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"strings"
+	"time"
+
+	"github.com/anthropics/agentsmesh/runner/internal/client"
+	"github.com/anthropics/agentsmesh/runner/internal/logger"
+	"github.com/anthropics/agentsmesh/runner/internal/terminal"
+	"github.com/anthropics/agentsmesh/runner/internal/terminal/aggregator"
+	"github.com/anthropics/agentsmesh/runner/internal/terminal/vt"
+)
+
+// Build creates the pod.
+func (b *PodBuilder) Build(ctx context.Context) (*Pod, error) {
+	if b.cmd == nil {
+		return nil, fmt.Errorf("command is required")
+	}
+	if b.cmd.PodKey == "" {
+		return nil, fmt.Errorf("pod key is required")
+	}
+	if b.cmd.LaunchCommand == "" {
+		return nil, fmt.Errorf("launch command is required")
+	}
+
+	logger.Pod().Info("Building pod", "pod_key", b.cmd.PodKey, "command", b.cmd.LaunchCommand)
+
+	// Report initial progress
+	b.sendProgress("pending", 0, "Initializing pod...")
+
+	// Setup sandbox and working directory
+	sandboxRoot, workingDir, branchName, err := b.setup(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	// Setup clipboard backend for image paste support
+	if sandboxRoot != "" && b.deps.Clipboard != nil {
+		if err := b.deps.Clipboard.Setup(sandboxRoot); err != nil {
+			logger.Pod().Warn("Failed to setup clipboard backend", "backend", b.deps.Clipboard.Name(), "error", err)
+			// Non-fatal: image paste won't work but pod still functions
+		}
+	}
+
+	// Resolve template variables in launch args
+	resolvedArgs := b.resolveArgs(b.cmd.LaunchArgs, sandboxRoot, workingDir)
+	logger.Pod().Debug("Resolved launch args", "pod_key", b.cmd.PodKey, "args", resolvedArgs)
+
+	// Merge environment variables
+	envVars := b.mergeEnvVars(sandboxRoot)
+	logger.Pod().Debug("Merged environment variables", "pod_key", b.cmd.PodKey, "count", len(envVars))
+
+	// Report progress: starting PTY
+	b.sendProgress("starting_pty", 80, "Starting terminal...")
+
+	// Create terminal
+	term, err := terminal.New(terminal.Options{
+		Command:  b.cmd.LaunchCommand,
+		Args:     resolvedArgs,
+		WorkDir:  workingDir,
+		Env:      envVars,
+		Rows:     b.rows,
+		Cols:     b.cols,
+		Label:    b.cmd.PodKey, // For log correlation in PTY diagnostics
+		OnOutput: nil,          // Will be wired up after all components are created
+		OnExit:   nil,          // Will be set by caller (MessageHandler)
+	})
+	if err != nil {
+		// Cleanup sandbox on failure
+		if sandboxRoot != "" {
+			os.RemoveAll(sandboxRoot)
+		}
+		return nil, &client.PodError{
+			Code:    client.ErrCodeCommandStart,
+			Message: fmt.Sprintf("failed to create terminal: %v", err),
+		}
+	}
+
+	// Create VirtualTerminal for terminal state management and snapshots
+	virtualTerm := vt.NewVirtualTerminal(b.cols, b.rows, b.vtHistoryLimit)
+	if b.oscHandler != nil {
+		virtualTerm.SetOSCHandler(b.oscHandler)
+	}
+
+	// Create SmartAggregator for adaptive frame rate output
+	agg := aggregator.NewSmartAggregator(nil, nil,
+		aggregator.WithFullRedrawThrottling(),
+	)
+
+	// Set up PTY logging if enabled
+	var ptyLogger *aggregator.PTYLogger
+	if b.enablePTYLogging && b.ptyLogDir != "" {
+		var logErr error
+		ptyLogger, logErr = aggregator.NewPTYLogger(b.ptyLogDir, b.cmd.PodKey)
+		if logErr != nil {
+			logger.Pod().Warn("Failed to create PTY logger", "pod_key", b.cmd.PodKey, "error", logErr)
+		} else {
+			agg.SetPTYLogger(ptyLogger)
+			logger.Pod().Info("PTY logging enabled for pod", "pod_key", b.cmd.PodKey, "log_dir", ptyLogger.LogDir())
+		}
+	}
+
+	// Create pod with all components
+	pod := &Pod{
+		ID:              b.cmd.PodKey,
+		PodKey:          b.cmd.PodKey,
+		AgentType:       "", // Could be extracted from command if needed
+		Branch:          branchName,
+		SandboxPath:     sandboxRoot,
+		Clipboard:       b.deps.Clipboard,
+		Terminal:        term,
+		VirtualTerminal: virtualTerm,
+		Aggregator:      agg,
+		PTYLogger:       ptyLogger,
+		StartedAt:       time.Now(),
+		Status:          PodStatusInitializing,
+	}
+
+	// Wire up output handler (integrates VirtualTerminal, StateDetector, and Aggregator)
+	podKey := b.cmd.PodKey
+	term.SetOutputHandler(func(data []byte) {
+		defer func() {
+			if r := recover(); r != nil {
+				logger.Terminal().Error("PANIC in OutputHandler recovered",
+					"pod_key", podKey,
+					"panic", fmt.Sprintf("%v", r),
+					"data_len", len(data))
+			}
+		}()
+
+		var screenLines []string
+		if virtualTerm != nil {
+			startFeed := time.Now()
+			screenLines = virtualTerm.Feed(data)
+			feedTime := time.Since(startFeed)
+			if feedTime > 100*time.Millisecond {
+				logger.Terminal().Warn("VT Feed slow",
+					"pod_key", podKey,
+					"data_len", len(data),
+					"feed_time", feedTime)
+			}
+		}
+		go pod.NotifyStateDetectorWithScreen(len(data), screenLines)
+
+		startWrite := time.Now()
+		agg.Write(data)
+		writeTime := time.Since(startWrite)
+		if writeTime > 100*time.Millisecond {
+			logger.Terminal().Warn("Aggregator Write slow",
+				"pod_key", podKey,
+				"data_len", len(data),
+				"write_time", writeTime)
+		}
+	})
+
+	logger.Pod().Info("Pod built", "pod_key", b.cmd.PodKey, "working_dir", workingDir, "cols", b.cols, "rows", b.rows)
+
+	// Report progress: ready
+	b.sendProgress("ready", 100, "Pod is ready")
+
+	return pod, nil
+}
+
+// resolvePath resolves path template variables.
+func (b *PodBuilder) resolvePath(pathTemplate, sandboxRoot, workDir string) string {
+	path := pathTemplate
+	path = strings.ReplaceAll(path, "{{.sandbox.root_path}}", sandboxRoot)
+	path = strings.ReplaceAll(path, "{{.sandbox.work_dir}}", workDir)
+	return path
+}
+
+// resolveArgs resolves template variables in command line arguments.
+func (b *PodBuilder) resolveArgs(args []string, sandboxRoot, workDir string) []string {
+	resolved := make([]string, len(args))
+	for i, arg := range args {
+		resolved[i] = b.resolvePath(arg, sandboxRoot, workDir)
+	}
+	return resolved
+}
+
+// mergeEnvVars merges all environment variable sources.
+func (b *PodBuilder) mergeEnvVars(sandboxRoot string) map[string]string {
+	result := make(map[string]string)
+
+	// Inject resolved login shell PATH as lowest-priority base.
+	// This ensures PTY processes can find tools (claude, aider, etc.)
+	// even when the runner runs as a systemd/launchd service with minimal PATH.
+	if b.deps.Config != nil && b.deps.Config.ResolvedPATH != "" {
+		result["PATH"] = b.deps.Config.ResolvedPATH
+	}
+
+	// Add config env vars (overrides resolved PATH if user sets PATH in config)
+	if b.deps.Config != nil {
+		for k, v := range b.deps.Config.AgentEnvVars {
+			result[k] = v
+		}
+	}
+
+	// Add command env vars (highest priority)
+	if b.cmd != nil {
+		for k, v := range b.cmd.EnvVars {
+			result[k] = v
+		}
+	}
+
+	// Apply clipboard backend env overrides (e.g., PATH prepend for shim)
+	if sandboxRoot != "" && b.deps.Clipboard != nil {
+		overrides := b.deps.Clipboard.EnvOverrides(sandboxRoot)
+		for k, v := range overrides {
+			if k == "PATH" {
+				// v is a directory to prepend to PATH
+				basePath := result["PATH"]
+				if basePath == "" {
+					basePath = os.Getenv("PATH")
+				}
+				if basePath == "" {
+					result["PATH"] = v
+				} else {
+					result["PATH"] = v + string(os.PathListSeparator) + basePath
+				}
+				continue
+			}
+			result[k] = v
+		}
+	}
+
+	return result
+}
+
+// sendProgress sends a pod initialization progress event to the server.
+// This is a best-effort operation - errors are logged but not returned.
+func (b *PodBuilder) sendProgress(phase string, progress int, message string) {
+	if b.cmd == nil || b.cmd.PodKey == "" || b.deps.ProgressSender == nil {
+		return
+	}
+
+	if err := b.deps.ProgressSender.SendPodInitProgress(b.cmd.PodKey, phase, int32(progress), message); err != nil {
+		logger.Pod().Debug("Failed to send init progress", "pod_key", b.cmd.PodKey, "phase", phase, "error", err)
+	}
+}
