+++ runner/internal/runner/pod_builder_integration_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package runner
 
 import (
