+++ runner/internal/terminal/aggregator/smart_aggregator_throttle_integration_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package aggregator
 
 import (
