+++ runner/internal/workspace/create_worktree_integration_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package workspace
 
 import (
